refactor: Move HarborAuthArgs and HarborProjectInfo to clients module
